feat: add new dependencies for third-party integration, refactor layout components to utilize dynamic settings, and enhance styles for improved responsiveness across various sections

+++ src/utils/helper.ts
@@ -2,7 +2,7 @@
 export function formatDate(date: string) {
   const options: Intl.DateTimeFormatOptions = {
     year: "numeric",
-    month: "short",
+    month: "numeric",
     day: "numeric",
   };
